Refactored Material: changed constructer and added attribute String description
