--- src/main/java/app/entities/MaterialHelper.java
@@ -1,26 +1,0 @@
-package app.entities;
-
-public class MaterialHelper
-{
-    /*
-    private String generateDescription()
-    {
-        StringBuilder description = new StringBuilder();
-        if (this.materialWidth != null && this.materialHeight != null)
-        {
-            description.append(materialWidth).append("x");
-            description.append(materialHeight).append(" mm. ");
-        }
-        else if (this.materialWidth != null && this.materialHeight == null)
-        {
-            description.append(materialWidth).append(" mm. ");
-        }
-        else if (this.materialWidth == null && this.materialHeight != null)
-        {
-            description.append(materialHeight).append(" mm. ");
-        }
-        description.append(name);
-        return description.toString();
-    }
-     */
-}
